requirements.txt
asgiref==3.8.1
djangO==4.2.28
djangorestframework==3.15.2
sqlparse==0.5.0
tzdata==2024.1
openai>=1.0.0
